refactor(utils): move change_field_type to pydantic utils file
