+++ liteswarm/utils/pydantic.py
@@ -16,12 +16,14 @@ from typing import (
     TypeGuard,
     TypeVar,
     Union,
+    Unpack,
     get_args,
     get_origin,
 )
 
 from pydantic import BaseModel, ValidationError, create_model
 from pydantic.fields import FieldInfo, _FromFieldInfoInputs
+from pydantic_core import PydanticUndefined
 
 T = TypeVar("T", bound=BaseModel)
 V = TypeVar("V", bound=BaseModel)
@@ -336,3 +338,107 @@ def restore_default_values(instance: T, target_model: type[V]) -> V:
         raise ValueError(
             f"Error restoring default values for model '{target_model.__name__}': {e}"
         ) from e
+
+
+def change_field_type(  # noqa: PLR0913
+    model_type: type[T],
+    field_name: str,
+    new_type: Any,
+    new_model_type: type[T] | None = None,
+    new_model_name: str | None = None,
+    default: Any = PydanticUndefined,
+    **kwargs: Unpack[_FromFieldInfoInputs],
+) -> type[T]:
+    """Create a new Pydantic model with a modified or added field.
+
+    Creates a copy of the original model with one field modified or added,
+    preserving all other fields and model configuration. The new model can:
+    - Modify existing field types
+    - Add new fields
+    - Change base model type
+    - Customize field validation
+
+    Args:
+        model_type: The original Pydantic model to modify
+        field_name: Name of the field to modify or add
+        new_type: New type for the field
+        new_model_type: Optional new base model type (defaults to original model)
+        new_model_name: Optional name for the new model (defaults to "Updated" + original name)
+        default: Optional default value for the field
+        **kwargs: Additional field configuration (validation rules, descriptions, etc.)
+
+    Returns:
+        A new Pydantic model class with the modified or added field
+
+    Raises:
+        TypeError: If the default value doesn't match the new field type
+        ValidationError: If default value validation fails when validate_default=True
+
+    Example:
+    ```python
+    class User(BaseModel):
+        id: int
+        name: str
+
+    # Modify existing field
+    UserStr = change_field_type(
+        model_type=User,
+        field_name="id",
+        new_type=str,
+        new_model_name="UserStr"
+    )
+
+    # Add new field
+    UserWithAge = change_field_type(
+        model_type=User,
+        field_name="age",
+        new_type=int,
+        default=0,
+        ge=0,
+        description="User's age in years"
+    )
+
+    # Change base model and add validation
+    class ValidatedModel(BaseModel):
+        model_config = ConfigDict(validate_default=True)
+
+    UserValidated = change_field_type(
+        model_type=User,
+        field_name="name",
+        new_type=str,
+        new_model_type=ValidatedModel,
+        min_length=1,
+        max_length=100
+    )
+    ```
+    """
+    fields: dict[str, Any] = {}
+    if field := model_type.model_fields.get(field_name):
+        field_info = copy_field_info(field, default=default, **kwargs)
+        fields[field_name] = (new_type, field_info)
+    else:
+        fields[field_name] = (new_type, FieldInfo(default=default, **kwargs))
+
+    for name, field in model_type.model_fields.items():
+        if name != field_name:
+            fields[name] = (field.annotation, field)
+
+    if default is not PydanticUndefined and kwargs.get("validate_default"):
+        try:
+            fields = {field_name: (new_type, default)}
+            temp_model = create_model("TempModel", **fields)
+            temp_model(**{field_name: default})
+        except ValidationError as e:
+            raise TypeError(
+                f"Default value {default!r} is not valid for field '{field_name}' of type {new_type}"
+            ) from e
+
+    updated_model_name = new_model_name or f"Updated{model_type.__name__}"
+
+    new_model = create_model(
+        updated_model_name,
+        __base__=new_model_type or model_type,
+        **fields,
+    )
+
+    return new_model
